// package/wagmiGenerated.ts
export const zoraCreator1155FactoryImplABI = [
  {
    stateMutability: 'nonpayable',
    type: 'constructor',
    inputs: [
      {
        name: '_implementation',
        internalType: 'contract IZoraCreator1155',
        type: 'address',
      },
      {
        name: '_merkleMinter',
        internalType: 'contract IMinter1155',
        type: 'address',
      },
      {
        name: '_fixedPriceMinter',
        internalType: 'contract IMinter1155',
        type: 'address',
      },
      {
        name: '_redeemMinterFactory',
        internalType: 'contract IMinter1155',
        type: 'address',
      },
    ],
  },
// ...
] as const
// ...
export const zoraCreator1155FactoryImplAddress = {
  1: '0xA6C5f2DE915240270DaC655152C3f6A91748cb85',
  5: '0x8732b4bCa198509bB9c40f9a24638Be1eaB7D30c',
  10: '0x78b524931e9d847c40BcBf225c25e154a7B05fDA',
  420: '0xb0C56317E9cEBc6E0f7A59458a83D0A9ccC3e955',
  999: '0x6a357139C1bcDcf0B3AB9bC447932dDdcb956703',
  11155111: '0x13dAA8E9e3f68deDE7b1386ACdc12eA98F2FB688',
} as const
// ...
export const zoraCreator1155FactoryImplConfig = {
  address: zoraCreator1155FactoryImplAddress,
  abi: zoraCreator1155FactoryImplABI,
} as const
// ...
export const zoraCreator1155PreminterABI = [
  {
    type: 'error',
    inputs: [],
    name: 'INITIALIZABLE_CONTRACT_ALREADY_INITIALIZED',
  },
// ...
] as const